requirements.txt
absl-py==0.9.0
altair==4.1.0
astor==0.8.1
attrs==19.3.0
backcall==0.2.0
base58==2.0.1
beautifulsoup4==4.6.0
billiard==3.3.0.23
bleach==3.1.5
blinker==1.4
boto3==1.14.13
botocore==1.17.13
cachetools==4.0.0
celery==3.1.18
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
configparser==4.0.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20121119
django-crispy-forms==1.7.2
django-decouple==2.1
django-extensions==2.0.7
django-import-export==1.2.0
django-jet==1.0.7
django-newsletter==0.6
django-parler==1.5.1
django-paypal==0.2.5
django-registration-redux==1.10
django-sendgrid-v5==0.6.87
django-tinymce==2.4.0
django-watson==1.5.2
django-webtest==1.9.2
djangorestframework==3.8.2
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
feedparser==5.2.1
google-api-python-client==1.4.1
google-auth==1.10.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
httplib2==0.10.3
idna==2.6
importlib-metadata==1.7.0
ipykernel==5.3.0
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.13.2
jsonschema==3.2.0
jupyter-client==6.1.5
jupyter-core==4.6.3
kiwisolver==1.1.0
kombu==3.0.37
ldif3==3.1.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.3
microsofttranslator==0.8
mistune==0.8.4
notebook==6.0.3
numexpr==2.7.0
numpy==1.17.5
oauth2client==2.2.0
odfpy==1.3.6
openpyxl==2.5.3
opt-einsum==3.1.0
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pathtools==0.1.2
pickleshare==0.7.5
Pillow==7.2.0
plotly==4.8.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.11.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pydeck==0.4.0b2
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-card-me==0.9.3
python-dateutil==2.8.1
python-http-client==3.0.0
pytz==2019.3
pywinpty==0.5.7
PyYAML==3.12
pyzmq==19.0.1
requests==2.22.0
retrying==1.3.3
rsa==3.4.2
s3transfer==0.3.3
scikit-learn==0.22.1
scipy==1.3.3
seaborn==0.10.1
Send2Trash==1.5.0
sendgrid==5.3.0
six==1.13.0
sorl-thumbnail==12.4.1
South==1.0.2
streamlit==0.62.1
surlex==0.2.0
tablib==0.12.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
toolz==0.10.0
tornado==6.0.4
traitlets==4.3.3
tzlocal==2.1
uritemplate==3.0.0
urllib3==1.22
validators==0.15.0
waitress==1.1.0
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.1
WebTest==2.0.29
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.11.2
zipp==3.1.0
